+++ src/apps/softwareupdater/UpdateManager.cpp
@@ -535,10 +535,8 @@ UpdateManager::_UpdateDownloadProgress(const char* header,
 		throw BAbortedByUserException();
 
 	BString packageCount;
-	packageCount.SetToFormat(
-		B_TRANSLATE_COMMENT("%i of %i", "Do not translate %i"),
-		fPackageDownloadsCount,
-		fPackageDownloadsTotal);
+	packageCount.SetToFormat(B_TRANSLATE_COMMENT("%i of %i", "Do not translate %i"),
+		(int)fPackageDownloadsCount, (int)fPackageDownloadsTotal);
 	BMessage message(kMsgProgressUpdate);
 	if (header != NULL)
 		message.AddString(kKeyHeader, header);
