+++ src/kits/tracker/WidgetAttributeText.cpp
@@ -96,8 +96,8 @@ TruncFileSizeBase(BString* outString, int64 value, const View* view,
 
 	if (value < kKBSize) {
 		if (view->StringWidth(buffer.String()) > width) {
-			buffer.SetToFormat(B_TRANSLATE_COMMENT("%lld B",
-				"The filesize symbol for byte"), value);
+			buffer.SetToFormat(B_TRANSLATE_COMMENT("%lld B", "The filesize symbol for byte"),
+				(long long int)value);
 		}
 	} else {
 		// strip off an insignificant zero so we don't get readings
